+++ benchmark/thermalstate/lowtemp_l10.jl
@@ -72,7 +72,7 @@ function bench_lowtemp()
     ψ = changebond!(vectorize(infinite_temperature_state(ComplexF64, ds)); D=32, noise=0)
     restore_gauge!(ψ)
     env = DMRGCache(superoperator(H, :left) + superoperator(H, :right), ψ)
-    alg = TDVP1(stepsize=-im * β / 2 / 20, verbosity=0)
+    alg = TDVP1(stepsize=-β / 2 / 20, verbosity=0)
     for _ in 1:20
         sweep!(env, alg)
     end